requirements.txt
apache-airflow
pandas
s3fs
boto3
requests